Fixed Connections Downloading works now need to implement blocks in pieces and implement endgame algo

+++ Networking/src/NetworkController.java
@@ -1,52 +1,83 @@
+import java.io.File;
+import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.ObjectOutputStream;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
 
 public class NetworkController {
     private static ArrayList<Torrent> torrents;
-    private static ArrayList<Connection> connections;
+    private static List<Connection> connections;
     private static Server server;
     private static Thread serverThread;
 
-    public static ArrayList<Connection> getConnections(){return connections;}
+    public static List<Connection> getConnections() {
+        return connections;
+    }
 
-    public static void startServer() throws IOException
-    {
+    public static void startServer() throws IOException {
         server = new Server();
         serverThread = new Thread(server);
         serverThread.setDaemon(true);
         serverThread.start();
     }
 
-    public static void killServer()
-    {
+    public static void killServer() {
 
-        server.kill();
+        if (server!=null)
+           server.kill();
 
     }
-
-    public static void addTorrent(Torrent t){torrents.add(t);}
-
-    public static void invokeTorrents()
-    {
+    public static void killTorrents(){
         for (Torrent t:torrents){
+            t.killThreads();
+        }
+        try {
+            File fl = new File("torrents.list");
+            fl.createNewFile();
+            ObjectOutputStream obs = new ObjectOutputStream(new FileOutputStream(fl));
+            obs.writeObject(torrents);
+        }catch (IOException ioe){}
+    }
+    public static void addTorrent(Torrent t) {
+        torrents.add(t);
+    }
+    public static void addTorrents(ArrayList<Torrent> ts){
+        torrents.addAll(ts);
+    }
+
+    public static void invokeTorrents() {
+        for (Torrent t : torrents) {
             t.invokeThreads();
         }
     }
 
-    public static ArrayList<Torrent> getTorrents(){return torrents;}
+    public static ArrayList<Torrent> getTorrents() {
+        return torrents;
+    }
 
-    public static boolean checkIfTorrentExists(byte arr[])
-    {
+    public static Torrent checkIfTorrentExists(byte arr[]) {
         for (Torrent t : torrents)
             if (Arrays.equals(t.getInfoHash(), arr))
-                return true;
+                return t;
 
+        return null;
+    }
+
+    public static boolean ipExists(String ip) {
+        synchronized(connections){
+           for (Connection c:connections){
+               if (c.getIP().equals(ip))
+                   return true;
+           }
+        }
         return false;
     }
 
     static {
         torrents = new ArrayList<>();
-        connections = new ArrayList<>();
+        connections = Collections.synchronizedList(new ArrayList<>());
     }
 }
